requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
crewai
crewai-tools
python-dotenv
asyncio
